Resize(TensorShape) for Tensor

+++ dali/pipeline/operators/detection/box_encoder.cc
@@ -165,7 +165,7 @@ void BoxEncoder<CPUBackend>::RunImpl(SampleWorkspace *ws, const int idx) {
   // Create output
   auto &bboxes_output = ws->Output<CPUBackend>(0);
   bboxes_output.set_type(bboxes_input.type());
-  bboxes_output.Resize({static_cast<int>(anchors_.size()), BoundingBox::kSize});
+  bboxes_output.Resize({static_cast<int>(anchors_.size()), static_cast<int>(BoundingBox::kSize)});
   auto out_boxes = bboxes_output.mutable_data<float>();
 
   auto &labels_output = ws->Output<CPUBackend>(1);
