+++ dali/pipeline/operators/detection/box_encoder.cu
@@ -32,8 +32,8 @@ void BoxEncoder<GPUBackend>::PrepareAnchors(const vector<float> &anchors) {
     "Anchors size must be divisible by 4, actual value = " + std::to_string(anchors.size()));
 
   anchors_count_ = anchors.size() / BoundingBox::kSize;
-  anchors_.Resize({anchors_count_, BoundingBox::kSize});
-  anchors_as_center_wh_.Resize({anchors_count_, BoundingBox::kSize});
+  anchors_.Resize({anchors_count_, static_cast<int64_t>(BoundingBox::kSize)});
+  anchors_as_center_wh_.Resize({anchors_count_, static_cast<int64_t>(BoundingBox::kSize)});
 
   auto anchors_data_cpu = reinterpret_cast<const float4 *>(anchors.data());
